requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.13.0
astunparse==1.6.3
audioread==3.1.0
av==17.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
ctranslate2==4.7.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
deepface==0.0.99
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
future==1.0.0
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.14.0
kiwisolver==1.5.0
lap==0.5.13
lazy-loader==0.5
libclang==18.1.1
librosa==0.11.0
lightdsa==0.0.3
lightecc==0.0.5
lightphe==0.0.23
llvmlite==0.47.0
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
namex==0.1.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.3
nvidia-cublas-cu12==12.9.2.10
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cudnn-cu12==9.20.0.48
ollama==0.6.1
onnxruntime-gpu==1.20.1
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
pillow==12.1.1
platformdirs==4.9.6
polars==1.39.3
polars-runtime-32==1.39.3
pooch==1.9.0
protobuf==7.34.1
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.20.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
retina-face==0.0.17
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
sympy==1.13.1
tensorflow==2.21.0
termcolor==3.3.0
tf_keras==2.21.0
threadpoolctl==3.6.0
tokenizers==0.21.4
# torch==2.5.1+cu121
# torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==4.49.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.36
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
